GetAllClasses generalized
