+++ WCF-Generator/Common/Infrastructure/ClassCompilerInfo.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.CodeAnalysis;
+﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace WCFGenerator.MappingsGeneration.Infrastructure
+namespace WCFGenerator.Common.Infrastructure
 {
     public class ClassCompilerInfo
     {
