+++ WCF-Generator/MappingsGeneration/MappingGenerator.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using WCFGenerator.Common;
+using WCFGenerator.Common.Infrastructure;
 using WCFGenerator.MappingsGeneration.Configuration;
 using WCFGenerator.MappingsGeneration.Infrastructure;
 using WCFGenerator.MappingsGenerator.Analysis;
